% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spccharter.R
\name{spccharter}
\alias{spccharter}
\title{spccharter}
\usage{
spccharter(
  df,
  numerator,
  denominator = NULL,
  datecol = NULL,
  by,
  plot_type = c("c", "p", "u"),
  runlength = 8,
  initial_rows = 25,
  look_forward = 25,
  direction = c("above", "below", "both"),
  round_digits = 2,
  multiplier = c(100, 1000, 10000),
  facet_cols = NULL,
  facet_scales = "fixed",
  chart_title = NULL,
  chart_subtitle = NULL,
  chart_caption = NULL,
  chart_breaks = NULL,
  line_colr = "#005EB8",
  line_size = 1.1,
  point_colr = "#005EB8",
  point_size = 2.7,
  centre_colr = "#06425AFF",
  centre_line_size = 1.05,
  highlight_fill = "#B50A2AFF",
  cl_fill = "grey80",
  cl_colr = NULL,
  wl_fill = "grey90",
  wl_colr = NULL,
  overwrite_theme = TRUE,
  ...
)
}
\arguments{
\item{df}{data.frame or data table}

\item{numerator}{numeric value representing the number of defect(s)}

\item{denominator}{sample size}

\item{datecol}{name of date column}

\item{by}{a single unquoted  variable or character vector oflength 2
indicating desired grouping variable(s) . You must supply a grouping variable .}

\item{plot_type}{'c', 'p' or 'u' chart}

\item{runlength}{length of desired run. Less than 8 may not be statisticall significant}

\item{initial_rows}{number of points to calculate initial baseline mean}

\item{look_forward}{number of rows to rebase limits on, including those in the sustained run}

\item{direction}{should run occur "above", "below" or on "both" sides of the mean line}

\item{round_digits}{the number of decimal places to round the p / u values and limits to}

\item{multiplier}{to express results as rate per 1000, 10000 or for percentages}

\item{facet_cols}{how many columns are required in the plot facets}

\item{facet_scales}{defaults to "fixed". Alternatively, "free_y"}

\item{chart_title}{title for the  final chart}

\item{chart_subtitle}{subtitle for chart}

\item{chart_caption}{caption for chart}

\item{chart_breaks}{character string defining desired x-axis date breaks}

\item{line_colr}{colour for basic chart lines}

\item{line_size}{thickness of connecting lines between run chart points}

\item{point_colr}{colour for basic chart points}

\item{point_size}{size of normal run chart points}

\item{centre_colr}{colour for solid and extended mean  lines}

\item{centre_line_size}{thickness of solid and extended mean  lines}

\item{highlight_fill}{fill colour for highlighting points in a sustained run}

\item{cl_fill}{geom_ribbon fill for upper  and lower control limits}

\item{cl_colr}{optional line colour for upper and lower control limits}

\item{wl_fill}{geom_ribbon fill for upper  and lower warning limits}

\item{wl_colr}{optional line colour for upper and lower warning limits}

\item{overwrite_theme}{set to FALSE if you want to amend the final plot
afterwards, in which case it returns the default ggplot2 theme, gridlines
and date labels. Leave at TRUE for theme_minimal, no gridlines and rotated
date labels.}

\item{...}{further arguments passed on to function}
}
\value{
(faceted) plot plus detailed data.tables showing improvement data.
These outputs can be used for further processing in external tools
}
\description{
Finds all runs of desired  length occurring on desired side of centre line.
Can also find runs occurring on both sides of the line, though this is of
limited use in terms of quality improvement.
Re-bases centre line each time a run is discovered.
The rebasing uses the points in the run, and a number of future points to ensure robust limits.
}
\details{
Facets and axis limits are handled by ggplot, though x-axis breaks can be
specified using the appropriate character string e.g. "3 months"
}
