% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{#' 75 grouped observations over time.}
\format{
A data frame with 75 rows and 4 variables:
\describe{
\item{report_month}{date of the observation, by month}
\item{defects}{integers representing counts of defects over time}
\item{possible}{integers representing the sample from which the defects occured}
\item{testgroup}{a grouping variable}

}
}
\usage{
testdata
}
\description{
A dataset containing 1 group of 75 integers showing counts of defects
from a sample size over several years.
}
\keyword{datasets}
